+++ src/lib/utils/plugins.ts
@@ -13,13 +13,25 @@ export async function loadPlugins(
         const pluginDisplay = getPluginDisplayName(plugin);
 
         try {
-            // On Windows, we need to ensure this path is a file path.
-            // Or we'll get ERR_UNSUPPORTED_ESM_URL_SCHEME
-            const esmPath = isAbsolute(plugin)
-                ? pathToFileURL(plugin).toString()
-                : plugin;
-            const instance = await import(esmPath);
-            const initFunction = instance.load ?? instance.default.load;
+            let instance: any;
+            // Try importing first to avoid warnings about requiring ESM being experimental.
+            // If that fails due to importing a directory, fall back to require.
+            try {
+                // On Windows, we need to ensure this path is a file path.
+                // Or we'll get ERR_UNSUPPORTED_ESM_URL_SCHEME
+                const esmPath = isAbsolute(plugin)
+                    ? pathToFileURL(plugin).toString()
+                    : plugin;
+                instance = await import(esmPath);
+            } catch (error: any) {
+                if (error.code === "ERR_UNSUPPORTED_DIR_IMPORT") {
+                    // eslint-disable-next-line @typescript-eslint/no-require-imports
+                    instance = require(plugin);
+                } else {
+                    throw error;
+                }
+            }
+            const initFunction = instance.load;
 
             if (typeof initFunction === "function") {
                 await initFunction(app);