+++ src/test/converter2/issues/gh2725.ts
@@ -1,0 +1,6 @@
+import { Node } from "typescript";
+
+/**
+ * {@link Node}
+ */
+export const node = true;